fix: bug in pagenation
Pagenation logic was if the new page is empty go back to prvious stage. This was preventing the last note to be deleted, as the page became empty so 2nd last state was being rendred. To fix this while changing page, if page is empty going to page-1 unless it is the 1st page